% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_matrix.R
\name{conf.matrix}
\alias{conf.matrix}
\title{conf.matrix}
\usage{
conf.matrix(p, x, tarvar = tarvar)
}
\arguments{
\item{p}{the probability which is the threshold of the pred.If pred exceeds,it will be assigned to 1.otherwise,it will be assigned to 0.}

\item{x}{the vector of the prediction of the xgboost model.}

\item{tarvar}{the target variable which is compared to the prediction.}
}
\value{
a confusion matrix
}
\description{
conf.matrix is a function which is used to generate a confusion matrix.
}
\examples{
conf.matrix(0.5,x=c(0:1),tarvar=c(0:1))
}
