% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{a simulation data}
\format{
An object of class matrix (inherits from array)
with 300 rows and 501 columns.
}
\usage{
dataset
}
\description{
A simulation data which is used to be an example
}
\keyword{datasets}
