% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redwine.R
\docType{data}
\name{redwine}
\alias{redwine}
\title{redwine quality data}
\format{
A data frame with 1599 rows and 12 variable
}
\usage{
redwine
}
\description{
A red wine quality data containing information of redwine
}
\keyword{datasets}
