% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_rmse.R
\name{xgb.rmse}
\alias{xgb.rmse}
\title{xgb.rmse}
\usage{
xgb.rmse(
  data,
  n,
  cvround = cvround,
  cvfold = cvfold,
  early_stopping_rounds = early_stopping_rounds,
  seed.number = seed.number,
  nthread = nthread
)
}
\arguments{
\item{data}{training dataset. xgb.rmse accepts only an \code{xgb.DMatrix} as the input.}

\item{n}{the number of cycles.}

\item{cvround}{max number of boosting iterations.}

\item{cvfold}{the original dataset is randomly partitioned into \code{nfold} equal size subsamples.}

\item{early_stopping_rounds}{If NULL, the early stopping function is not triggered. If set to an integer k,
training with a validation set will stop if the performance doesn't improve for k rounds.}

\item{seed.number}{random number seed.}

\item{nthread}{number of thread used in training.}
}
\value{
a \code{list} contains the best_param,the best_rmse and best_rmse_index.
}
\description{
xgb.rmse is an automatic parameter adjustment function which is applied to the regression task by xgboost.
}
\examples{

library(xgboost)
data(agaricus.train, package='xgboost')
data(agaricus.test, package='xgboost')
dtrain <- xgb.DMatrix(agaricus.train$data, label = agaricus.train$label)
dtest <- xgb.DMatrix(agaricus.test$data, label = agaricus.test$label)
fit <-xgb.rmse(dtrain,10,cvround=100,cvfold=5,
                  early_stopping_rounds =10,
                  seed.number = 12345,nthread = 8)
}
